% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationTable.R
\name{generateEntrezIdLinks}
\alias{generateEntrezIdLinks}
\title{Transform an ENTREZ ID into a hyperlink}
\usage{
generateEntrezIdLinks(x)
}
\arguments{
\item{x}{vector of Entrez IDs}
}
\value{
vector of hyperlinks for the corresponding Entrez IDs
}
\description{
Utility function to transform an ENTREZ ID into a hyperlink on 
the NCBI Entrez page for the given gene
}
\note{
Snippet taken from the \code{annaffy} package
}
\examples{
generateEntrezIdLinks(c("5230", "18655"))
}
\seealso{
\code{\link{generateGOIdLinks}}
}
\author{
Tobias Verbeke
}
\keyword{manip}
