% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationTable.R
\name{generateGOIdLinks}
\alias{generateGOIdLinks}
\title{Transform a GO ID into a hyperlink}
\usage{
generateGOIdLinks(x)
}
\arguments{
\item{x}{vector of GO IDs}
}
\value{
vector of hyperlinks
}
\description{
Utility function to transform a GO ID into a hyperlink
to the corresponding page on the gene ontology website
}
\note{
Snippet taken from the \code{annaffy} package
}
\seealso{
\code{\link{generateEntrezIdLinks}}
}
\author{
Tobias Verbeke
}
\keyword{manip}
