% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationTable.R
\name{print.xtableAnnotationTable}
\alias{print.xtableAnnotationTable}
\title{Print method for 'xtableAnnotationTable' objects}
\usage{
\method{print}{xtableAnnotationTable}(x, ...)
}
\arguments{
\item{x}{Object of class 'xtableAnnotationTable'}

\item{...}{Further arguments passed to \code{print.xtable}}
}
\value{
No returned value, the object is printed.
}
\description{
Print method for 'xtableAnnotationTable' objects
}
\details{
Wrapper to be able to use a specific \code{sanitize.text} function
 in the \code{print.xtable} call
}
\seealso{
\code{\link[xtable]{print.xtable}}
}
\author{
Tobias Verbeke
}
\keyword{manip}
